+++ src/core/objects/stepper.h
@@ -38,7 +38,7 @@ struct stepper_enable_args {
 };
 
 struct stepper_set_speed_args {
-    float steps_per_second;
+    double steps_per_second;
 };
 
 struct stepper_move_args {
